inventario-retail/requirements.txt
# Sistema de Inventario Multi-Agente - Argentina
# Post-MVP con ML, UI y Enhanced Dashboard
# ==================================================

# FastAPI y servidor web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Base de datos y ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Pydantic para validación
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2

# OCR y procesamiento de imágenes
easyocr==1.7.0
pillow==10.1.0
opencv-python==4.8.1.78

# Streamlit para UI
streamlit==1.28.2
plotly==5.17.0

# Utilidades y logging
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Linting y formateo
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Producción
gunicorn==21.2.0
redis==5.0.1

# Específicos para Argentina
holidays==0.37  # Para feriados argentinos
locale==0.1.1  # Para formatos locales

# Monitoreo y métricas
prometheus-client==0.19.0

# Email y notificaciones
smtplib==3.11.6  # Built-in, versión específica para compatibilidad

# Procesamiento de archivos
openpyxl==3.1.2
xlsxwriter==3.1.9

# Networking y HTTP
aiohttp==3.9.1
websockets==12.0

# Charts y visualización (para dashboard)
matplotlib==3.8.2  # Backend para gráficos si es necesario
seaborn==0.13.0    # Para análisis estadístico

# Utilidades adicionales
click==8.1.7       # Para CLI tools
tqdm==4.66.1       # Progress bars
python-dotenv==1.0.0  # Variables de entorno
